added park/event/attraction to itinerary event listener
