+++ scripts/itinerary/ItinerarySelect.js
@@ -1,23 +1,23 @@
 import { saveItinerary } from "./ItineraryProvider.js"
 
 // const contentTarget = document.querySelector(".itinerary__list")
-const eventHub = document.querySelector(".container")
+// const eventHub = document.querySelector(".container")
 
-
-eventHub.addEventListener("click", clickEvent => {
-    if (clickEvent.target.id === "showItinerary") {
-        clickEvent.preventDefault()
-        const attraction = document.getElementById("itinerary__attraction").value
-        const eatery = document.getElementById(".itinerary__eatery").value
-        const park = document.getElementById("itinerary__park").value
-      // Make a new object representation of a note
-        const newItinerary = {
-        // Key/value pairs here
-        "park": park,
-        "eatery": eatery,
-        "attraction": attraction,
-        }
-      // Change API state and application state
-        saveItinerary(newItinerary)
-    }
-})
+ //COME BACK TO THIS 
+// eventHub.addEventListener("click", clickEvent => {
+//     if (clickEvent.target.id === "showItinerary") {
+//         clickEvent.preventDefault()
+//         const attraction = document.getElementById("itinerary__attraction").value
+//         const eatery = document.getElementById(".itinerary__eatery").value
+//         const park = document.getElementById("itinerary__park").value
+//       // Make a new object representation of a note
+//         const newItinerary = {
+//         // Key/value pairs here
+//         "park": park,
+//         "eatery": eatery,
+//         "attraction": attraction,
+//         }
+//       // Change API state and application state
+//         saveItinerary(newItinerary)
+//     }
+// })
